# internal/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Lurkarr API
  description: >-
    REST API for Lurkarr — an automated media management tool that integrates
    with Sonarr, Radarr, Lidarr, Readarr, Whisparr, Eros, Prowlarr, SABnzbd,
    and Seerr.
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /
    description: Current server
...
paths:
...
  /api/auth/login:
    post:
      operationId: login
      tags: [Auth]
      summary: Log in
      description: Authenticate with username/password. Returns session cookie and CSRF token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          headers:
            Set-Cookie:
              schema:
                type: string
        '401':
          description: Invalid credentials or TOTP required
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Error'
                  - $ref: '#/components/schemas/TOTPRequiredResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/auth/logout:
    post:
      operationId: logout
      tags: [Auth]
      summary: Log out
      description: Invalidate current session.
      security:
        - cookieAuth: []
      responses:
        '200':
          $ref: '#/components/responses/StatusOk'
  /api/auth/setup:
    post:
      operationId: setup
      tags: [Auth]
      summary: First-run setup
      description: Create initial admin user. Only works when no users exist.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetupRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '409':
          description: Setup already completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/auth/2fa/enable:
    post:
      operationId: enable2FA
      tags: [Auth]
      summary: Enable 2FA
      description: Generate TOTP secret and QR code for the current user.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: TOTP secret generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enable2FAResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/2fa/disable:
    post:
      operationId: disable2FA
      tags: [Auth]
      summary: Disable 2FA
      security:
        - cookieAuth: []
      responses:
        '200':
          $ref: '#/components/responses/StatusOk'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/auth/2fa/verify:
    post:
      operationId: verify2FA
      tags: [Auth]
      summary: Verify TOTP code
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        '200':
          description: Code verified
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    const: verified
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/auth/oidc/callback:
    get:
      operationId: oidcCallback
      tags: [Auth]
      summary: OIDC callback
      description: Handles the authorization code callback from the OIDC provider.
      parameters:
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        '302':
          description: Redirect to application after successful authentication

  # ── User ──────────────────────────────────────────────────────────────
  /api/user:
    get:
      operationId: getUser
      tags: [User]
      summary: Get current user profile
      security:
        - cookieAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/user/username:
    post:
      operationId: updateUsername
      tags: [User]
      summary: Update username
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/StatusOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/user/password:
    post:
      operationId: updatePassword
      tags: [User]
      summary: Update password
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/StatusOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── Instances ─────────────────────────────────────────────────────────
  /api/instances/{app}:
    get:
      operationId: listInstances
      tags: [Instances]
      summary: List instances for an app type
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/AppTypePath'
      responses:
        '200':
          description: List of instances (API keys masked)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppInstance'
        '400':
          $ref: '#/components/responses/BadRequest'

    post:
      operationId: createInstance
      tags: [Instances]
      summary: Create a new instance
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/AppTypePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstanceRequest'
      responses:
        '201':
          description: Instance created (API key masked)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppInstance'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/instances/{id}:
    put:
      operationId: updateInstance
      tags: [Instances]
      summary: Update an instance
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UUIDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInstanceRequest'
      responses:
        '200':
          $ref: '#/components/responses/StatusOk'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteInstance
      tags: [Instances]
      summary: Delete an instance
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UUIDPath'
      responses:
        '200':
          $ref: '#/components/responses/StatusOk'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/instances/test:
    post:
      operationId: testInstanceConnection
      tags: [Instances]
      summary: Test connection to an arr instance
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestInstanceRequest'
      responses:
        '200':
          description: Connection successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TestConnectionResult'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          description: SSRF protection — private URL blocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '502':
          $ref: '#/components/responses/BadGateway'

  # ── Settings ──────────────────────────────────────────────────────────
  /api/settings/general:
    get:
      operationId: getGeneralSettings
      tags: [Settings]
      summary: Get general settings
      description: Secret key is masked in the response.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: General settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralSettings'
    put:
      operationId: updateGeneralSettings
      tags: [Settings]
      summary: Update general settings
      description: Secret key is always preserved server-side and cannot be changed via API.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GeneralSettings'
      responses:
        '200':
          description: Updated settings (secret key masked)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralSettings'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/settings/{app}:
    get:
      operationId: getAppSettings
      tags: [Settings]
      summary: Get per-app settings
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/AppTypePath'
      responses:
        '200':
          description: App settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSettings'
        '400':
          $ref: '#/components/responses/BadRequest'

    put:
      operationId: updateAppSettings
      tags: [Settings]
      summary: Update per-app settings
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/AppTypePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppSettings'
      responses:
        '200':
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppSettings'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ── History ───────────────────────────────────────────────────────────
  /api/history:
    get:
      operationId: listHistory
      tags: [History]
      summary: List lurk history
      security:
        - cookieAuth: []
      parameters:
        - name: app
          in: query
          schema:
            $ref: '#/components/schemas/AppType'
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Paginated history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryResponse'
  /api/history/{app}:
    delete:
      operationId: deleteHistory
      tags: [History]
      summary: Delete history for an app type
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/AppTypePath'
      responses:
        '200':
          $ref: '#/components/responses/StatusOk'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ── Logs ──────────────────────────────────────────────────────────────
...
  /api/stats:
    get:
      operationId: getStats
      tags: [Stats]
      summary: Get lurk statistics
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Per-app, per-instance stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LurkStats'
  /api/stats/reset:
    post:
      operationId: resetStats
      tags: [Stats]
      summary: Reset all statistics
      security:
        - cookieAuth: []
      responses:
        '200':
          $ref: '#/components/responses/StatusOk'
  /api/stats/hourly-caps:
    get:
      operationId: getHourlyCaps
      tags: [Stats]
      summary: Get hourly API cap usage
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Hourly cap entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HourlyCap'

  # ── State ─────────────────────────────────────────────────────────────
  /api/state:
    get:
      operationId: getState
      tags: [State]
      summary: Get instance state and reset times
      security:
        - cookieAuth: []
      parameters:
        - name: app
          in: query
          schema:
            $ref: '#/components/schemas/AppType'
      responses:
        '200':
          description: Instance state entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StateEntry'
  /api/state/reset:
    post:
      operationId: resetState
      tags: [State]
      summary: Reset state for a specific instance
      security:
        - cookieAuth: []
      parameters:
        - name: app
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/AppType'
        - name: instance_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          $ref: '#/components/responses/StatusOk'
        '400':
          $ref: '#/components/responses/BadRequest'

  # ── Schedules ─────────────────────────────────────────────────────────
  /api/schedules:
    get:
      operationId: listSchedules
      tags: [Schedules]
      summary: List all schedules
      security:
        - cookieAuth: []
      responses:
        '200':
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Schedule'

    post:
      operationId: createSchedule
      tags: [Schedules]
      summary: Create a schedule
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
      responses:
        '201':
          description: Schedule created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/schedules/{id}:
    put:
      operationId: updateSchedule
      tags: [Schedules]
      summary: Update a schedule
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UUIDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
      responses:
        '200':
          description: Schedule updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '400':
          $ref: '#/components/responses/BadRequest'

    delete:
      operationId: deleteSchedule
      tags: [Schedules]
      summary: Delete a schedule
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/UUIDPath'
      responses:
        '200':
          $ref: '#/components/responses/StatusOk'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/schedules/history:
    get:
      operationId: getScheduleHistory
      tags: [Schedules]
      summary: Get schedule execution history
      security:
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        '200':
          description: Execution history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ScheduleExecution'

  # ── Prowlarr ──────────────────────────────────────────────────────────
  /api/prowlarr/settings:
    get:
      operationId: getProwlarrSettings
      tags: [Prowlarr]
      summary: Get Prowlarr settings
      description: API key is masked in the response.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Prowlarr settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProwlarrSettings'

    put:
      operationId: updateProwlarrSettings
      tags: [Prowlarr]
      summary: Update Prowlarr settings
      description: Send masked API key to preserve existing value.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProwlarrSettings'
      responses:
        '200':
          description: Updated settings (API key masked)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProwlarrSettings'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/prowlarr/indexers:
    get:
      operationId: getProwlarrIndexers
      tags: [Prowlarr]
      summary: Get Prowlarr indexers
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Indexer list (pass-through from Prowlarr API)
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '502':
          $ref: '#/components/responses/BadGateway'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: lurkarr_session
...
  responses:
    StatusOk:
      description: Success
      content:
        application/json:
          schema:
            type: object
            properties:
              status:
                type: string
                const: ok
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
        totp_code:
          type: string
    LoginResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        csrf_token:
          type: string
    TOTPRequiredResponse:
      type: object
      properties:
        error:
          type: string
          const: totp_required
        totp_required:
          type: boolean
          const: true
    SetupRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    Enable2FAResponse:
      type: object
      properties:
        secret:
          type: string
        qr_base64:
          type: string

    # ── User ────────────────────────────────────────────────────────────
...