# internal/server/server.go
package server

import (
	"context"
	"io/fs"
	"log/slog"
	"net"
	"net/http"
	"strings"
	"time"

	"github.com/gorilla/csrf"
	"github.com/prometheus/client_golang/prometheus/promhttp"
	"golang.org/x/time/rate"

	"github.com/lusoris/lurkarr/internal/api"
	"github.com/lusoris/lurkarr/internal/auth"
	"github.com/lusoris/lurkarr/internal/database"
	"github.com/lusoris/lurkarr/internal/middleware"
	"github.com/lusoris/lurkarr/internal/notifications"
	"github.com/lusoris/lurkarr/internal/scheduler"
)

// scalarHTML is the Scalar API reference page.
var scalarHTML = []byte(`<!doctype html>
<html>
<head>
  <title>Lurkarr API Reference</title>
  <meta charset="utf-8" />
  <meta name="viewport" content="width=device-width, initial-scale=1" />
</head>
<body>
  <script id="api-reference" data-url="spec"></script>
  <script src="https://cdn.jsdelivr.net/npm/@scalar/api-reference"></script>
</body>
</html>
`)

// Config holds server configuration.
type Config struct {
	Addr           string
	CSRFKey        []byte
	AllowedOrigins []string
	ProxyAuth      bool
	ProxyHeaders   []string
	TrustedProxies []*net.IPNet
	SecureCookie   bool
	BasePath       string
	OpenAPISpec    []byte
	FrontendFS     fs.FS

	// OIDC
	OIDCEnabled      bool
	OIDCIssuerURL    string
	OIDCClientID     string
	OIDCClientSecret string
	OIDCRedirectURL  string
	OIDCScopes       []string
	OIDCAutoCreate   bool
	OIDCAdminGroup   string
}

// Server is the main HTTP server.
type Server struct {
	httpServer *http.Server
}

// csrfInjectToken wraps a handler to expose the CSRF token in a response header
// so the SPA can read it and send it back on mutating requests.
func csrfInjectToken(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("X-CSRF-Token", csrf.Token(r))
		next.ServeHTTP(w, r)
	})
}

// spaHandler serves an SPA from an fs.FS. It tries the exact path first,
// then falls back to index.html for client-side routing.
func spaHandler(fsys fs.FS) http.Handler {
	fileServer := http.FileServerFS(fsys)
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		// Try to serve the exact file.
		path := strings.TrimPrefix(r.URL.Path, "/")
		if path == "" {
			path = "index.html"
		}
		if _, err := fs.Stat(fsys, path); err == nil {
			// Cache immutable assets aggressively.
			if strings.HasPrefix(path, "_app/immutable/") {
				w.Header().Set("Cache-Control", "public, max-age=31536000, immutable")
			}
			fileServer.ServeHTTP(w, r)
			return
		}
		// SPA fallback: serve index.html for all other paths.
		r.URL.Path = "/"
		fileServer.ServeHTTP(w, r)
	})
}

// New creates a new Server with all routes registered.
func New(cfg Config, db *database.DB, sched *scheduler.Scheduler, notifMgr *notifications.Manager) *Server {
	// Set trusted proxies for rate limiter IP extraction.
	middleware.TrustedProxies = cfg.TrustedProxies

	authMw := &auth.Middleware{
		DB:              db,
		ProxyAuthBypass: cfg.ProxyAuth,
		ProxyHeaders:    cfg.ProxyHeaders,
		TrustedProxies:  cfg.TrustedProxies,
		ProxyAutoCreate: cfg.ProxyAuth,
		CSRFKey:         cfg.CSRFKey,
		SecureCookie:    cfg.SecureCookie,
	}

	authH := &api.AuthHandler{DB: db, Auth: authMw}
	settingsH := &api.SettingsHandler{DB: db}
	appsH := &api.AppsHandler{DB: db}
	historyH := &api.HistoryHandler{DB: db}
	statsH := &api.StatsHandler{DB: db}
	stateH := &api.StateHandler{DB: db}
	schedulerH := &api.SchedulerHandler{DB: db, Scheduler: sched}
	prowlarrH := &api.ProwlarrHandler{DB: db}
	sabnzbdH := &api.SABnzbdHandler{DB: db}
	userH := &api.UserHandler{DB: db}
	queueH := &api.QueueHandler{DB: db}
	blocklistH := &api.BlocklistHandler{DB: db}
	notificationH := &api.NotificationHandler{DB: db, Manager: notifMgr}
	seerrH := &api.SeerrHandler{DB: db}
	dlClientH := &api.DownloadClientHandler{DB: db}

	mux := http.NewServeMux()

	// Rate limiter for login: 5 attempts per minute per IP (burst 5).
	loginRL := middleware.NewIPRateLimiter(rate.Limit(5.0/60.0), 5)

	// --- Public routes (no auth) ---
	mux.Handle("POST /api/auth/login", middleware.RateLimit(loginRL)(http.HandlerFunc(authH.HandleLogin)))
	mux.HandleFunc("GET /api/auth/setup", authH.HandleSetupCheck)
	mux.HandleFunc("POST /api/auth/setup", authH.HandleSetup)

	// --- OIDC routes (public, no auth) ---
	if cfg.OIDCEnabled {
		oidcH := auth.NewOIDCHandler(auth.OIDCConfig{
			Enabled:      cfg.OIDCEnabled,
			IssuerURL:    cfg.OIDCIssuerURL,
			ClientID:     cfg.OIDCClientID,
			ClientSecret: cfg.OIDCClientSecret,
			RedirectURL:  cfg.OIDCRedirectURL,
			Scopes:       cfg.OIDCScopes,
			AutoCreate:   cfg.OIDCAutoCreate,
			AdminGroup:   cfg.OIDCAdminGroup,
		}, db, authMw)
		mux.HandleFunc("GET /api/auth/oidc/login", oidcH.HandleLogin)
		mux.HandleFunc("GET /api/auth/oidc/callback", oidcH.HandleCallback)
	}
	mux.HandleFunc("GET /api/auth/oidc/info", func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Content-Type", "application/json")
		if cfg.OIDCEnabled {
			_, _ = w.Write([]byte(`{"enabled":true}`))
		} else {
			_, _ = w.Write([]byte(`{"enabled":false}`))
		}
	})

	// --- Kubernetes health probes ---
	mux.HandleFunc("GET /healthz", func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Content-Type", "application/json")
		_, _ = w.Write([]byte(`{"status":"ok"}`))
	})
	mux.HandleFunc("GET /readyz", func(w http.ResponseWriter, r *http.Request) {
		if err := db.HealthCheck(r.Context()); err != nil {
			http.Error(w, `{"status":"not ready"}`, http.StatusServiceUnavailable)
			return
		}
		w.Header().Set("Content-Type", "application/json")
		_, _ = w.Write([]byte(`{"status":"ok"}`))
	})

	// --- Metrics (Prometheus) ---
	mux.Handle("GET /metrics", promhttp.Handler())

	// --- API Documentation ---
	if len(cfg.OpenAPISpec) > 0 {
		mux.HandleFunc("GET /api/spec", func(w http.ResponseWriter, r *http.Request) {
			w.Header().Set("Content-Type", "application/yaml")
			w.Header().Set("Cache-Control", "public, max-age=3600")
			_, _ = w.Write(cfg.OpenAPISpec)
		})
		mux.HandleFunc("GET /api/docs", func(w http.ResponseWriter, r *http.Request) {
			w.Header().Set("Content-Type", "text/html; charset=utf-8")
			_, _ = w.Write(scalarHTML)
		})
	}

	// --- Protected routes ---
	protected := http.NewServeMux()

	// Auth
	protected.HandleFunc("POST /api/auth/logout", authH.HandleLogout)
	protected.HandleFunc("POST /api/auth/2fa/enable", authH.Handle2FAEnable)
	protected.HandleFunc("POST /api/auth/2fa/disable", authH.Handle2FADisable)
	protected.HandleFunc("POST /api/auth/2fa/verify", authH.Handle2FAVerify)

	// User
	protected.HandleFunc("GET /api/user", userH.HandleGetUser)
	protected.HandleFunc("POST /api/user/username", userH.HandleUpdateUsername)
	protected.HandleFunc("POST /api/user/password", userH.HandleUpdatePassword)

	// Settings
	protected.HandleFunc("GET /api/settings/general", settingsH.HandleGetGeneralSettings)
	protected.HandleFunc("PUT /api/settings/general", settingsH.HandleUpdateGeneralSettings)
	protected.HandleFunc("GET /api/settings/{app}", settingsH.HandleGetAppSettings)
	protected.HandleFunc("PUT /api/settings/{app}", settingsH.HandleUpdateAppSettings)

	// App Instances
	protected.HandleFunc("GET /api/instances/{app}", appsH.HandleListInstances)
	protected.HandleFunc("POST /api/instances/{app}", appsH.HandleCreateInstance)
	protected.HandleFunc("PUT /api/instances/{id}", appsH.HandleUpdateInstance)
	protected.HandleFunc("DELETE /api/instances/{id}", appsH.HandleDeleteInstance)
	protected.HandleFunc("GET /api/instances/{id}/health", appsH.HandleHealthCheckInstance)
	protected.HandleFunc("POST /api/instances/test", appsH.HandleTestConnection)

	// History
	protected.HandleFunc("GET /api/history", historyH.HandleListHistory)
	protected.HandleFunc("DELETE /api/history/{app}", historyH.HandleDeleteHistory)

	// Stats
	protected.HandleFunc("GET /api/stats", statsH.HandleGetStats)
	protected.HandleFunc("POST /api/stats/reset", statsH.HandleResetStats)
	protected.HandleFunc("GET /api/stats/hourly-caps", statsH.HandleGetHourlyCaps)

	// State
	protected.HandleFunc("GET /api/state", stateH.HandleGetState)
	protected.HandleFunc("POST /api/state/reset", stateH.HandleResetState)

	// Schedules
	protected.HandleFunc("GET /api/schedules", schedulerH.HandleListSchedules)
	protected.HandleFunc("POST /api/schedules", schedulerH.HandleCreateSchedule)
	protected.HandleFunc("PUT /api/schedules/{id}", schedulerH.HandleUpdateSchedule)
	protected.HandleFunc("DELETE /api/schedules/{id}", schedulerH.HandleDeleteSchedule)
	protected.HandleFunc("GET /api/schedules/history", schedulerH.HandleScheduleHistory)

	// Prowlarr
	protected.HandleFunc("GET /api/prowlarr/settings", prowlarrH.HandleGetSettings)
	protected.HandleFunc("PUT /api/prowlarr/settings", prowlarrH.HandleUpdateSettings)
	protected.HandleFunc("GET /api/prowlarr/indexers", prowlarrH.HandleGetIndexers)
	protected.HandleFunc("GET /api/prowlarr/indexers/stats", prowlarrH.HandleGetIndexerStats)
	protected.HandleFunc("POST /api/prowlarr/test", prowlarrH.HandleTestConnection)

	// SABnzbd
	protected.HandleFunc("GET /api/sabnzbd/settings", sabnzbdH.HandleGetSettings)
	protected.HandleFunc("PUT /api/sabnzbd/settings", sabnzbdH.HandleUpdateSettings)
	protected.HandleFunc("GET /api/sabnzbd/queue", sabnzbdH.HandleGetQueue)
	protected.HandleFunc("GET /api/sabnzbd/history", sabnzbdH.HandleGetHistory)
	protected.HandleFunc("GET /api/sabnzbd/stats", sabnzbdH.HandleGetStats)
	protected.HandleFunc("POST /api/sabnzbd/pause", sabnzbdH.HandlePause)
	protected.HandleFunc("POST /api/sabnzbd/resume", sabnzbdH.HandleResume)
	protected.HandleFunc("POST /api/sabnzbd/test", sabnzbdH.HandleTestConnection)

	// Queue Management
	protected.HandleFunc("GET /api/queue/settings/{app}", queueH.HandleGetQueueCleanerSettings)
	protected.HandleFunc("PUT /api/queue/settings/{app}", queueH.HandleUpdateQueueCleanerSettings)
	protected.HandleFunc("GET /api/queue/scoring/{app}", queueH.HandleGetScoringProfile)
	protected.HandleFunc("PUT /api/queue/scoring/{app}", queueH.HandleUpdateScoringProfile)
	protected.HandleFunc("GET /api/queue/blocklist/{app}", queueH.HandleGetBlocklistLog)
	protected.HandleFunc("GET /api/queue/imports/{app}", queueH.HandleGetAutoImportLog)
	protected.HandleFunc("GET /api/queue/download-client/{app}", queueH.HandleGetDownloadClientSettings)
	protected.HandleFunc("PUT /api/queue/download-client/{app}", queueH.HandleUpdateDownloadClientSettings)

	// Blocklist Sources & Rules
	protected.HandleFunc("GET /api/blocklist/sources", blocklistH.HandleListSources)
	protected.HandleFunc("POST /api/blocklist/sources", blocklistH.HandleCreateSource)
	protected.HandleFunc("GET /api/blocklist/sources/{id}", blocklistH.HandleGetSource)
	protected.HandleFunc("PUT /api/blocklist/sources/{id}", blocklistH.HandleUpdateSource)
	protected.HandleFunc("DELETE /api/blocklist/sources/{id}", blocklistH.HandleDeleteSource)
	protected.HandleFunc("GET /api/blocklist/rules", blocklistH.HandleListRules)
	protected.HandleFunc("POST /api/blocklist/rules", blocklistH.HandleCreateRule)
	protected.HandleFunc("DELETE /api/blocklist/rules/{id}", blocklistH.HandleDeleteRule)

	// Notifications
	protected.HandleFunc("GET /api/notifications/providers", notificationH.HandleListProviders)
	protected.HandleFunc("POST /api/notifications/providers", notificationH.HandleCreateProvider)
	protected.HandleFunc("GET /api/notifications/providers/{id}", notificationH.HandleGetProvider)
	protected.HandleFunc("PUT /api/notifications/providers/{id}", notificationH.HandleUpdateProvider)
	protected.HandleFunc("DELETE /api/notifications/providers/{id}", notificationH.HandleDeleteProvider)
	protected.HandleFunc("POST /api/notifications/providers/{id}/test", notificationH.HandleTestProvider)

	// Download Client Instances
	protected.HandleFunc("GET /api/download-clients", dlClientH.HandleList)
	protected.HandleFunc("POST /api/download-clients", dlClientH.HandleCreate)
	protected.HandleFunc("PUT /api/download-clients/{id}", dlClientH.HandleUpdate)
	protected.HandleFunc("DELETE /api/download-clients/{id}", dlClientH.HandleDelete)
	protected.HandleFunc("GET /api/download-clients/{id}/health", dlClientH.HandleHealthCheck)
	protected.HandleFunc("POST /api/download-clients/test", dlClientH.HandleTest)

	// Seerr
	protected.HandleFunc("GET /api/seerr/settings", seerrH.HandleGetSettings)
	protected.HandleFunc("PUT /api/seerr/settings", seerrH.HandleUpdateSettings)
	protected.HandleFunc("POST /api/seerr/test", seerrH.HandleTestConnection)
	protected.HandleFunc("GET /api/seerr/requests", seerrH.HandleGetRequests)
	protected.HandleFunc("GET /api/seerr/requests/count", seerrH.HandleGetRequestCount)

	// Mount protected routes with auth + CSRF + token injection
	mux.Handle("/api/", authMw.CSRFProtect()(csrfInjectToken(authMw.RequireAuth(protected))))

	// Serve embedded frontend SPA (if built).
	if cfg.FrontendFS != nil {
		mux.Handle("/", spaHandler(cfg.FrontendFS))
	}

	// Global middleware chain
	corsMiddleware := middleware.CORS(middleware.CORSConfig{AllowedOrigins: cfg.AllowedOrigins})
	var handler http.Handler
	handler = middleware.Chain(mux,
		middleware.Recovery,
		middleware.RequestID,
		middleware.Logging,
		corsMiddleware,
	)

	// Strip base path prefix if configured (for sub-path reverse proxy hosting).
	if cfg.BasePath != "" {
		handler = http.StripPrefix(cfg.BasePath, handler)
		slog.Info("base path configured", "base_path", cfg.BasePath)
	}

	return &Server{
		httpServer: &http.Server{
			Addr:              cfg.Addr,
			Handler:           handler,
			ReadHeaderTimeout: 10 * time.Second,
			IdleTimeout:       120 * time.Second,
		},
	}
}

// Start begins listening and serving.
func (s *Server) Start() error {
	slog.Info("server starting", "addr", s.httpServer.Addr)
	if err := s.httpServer.ListenAndServe(); err != nil && err != http.ErrServerClosed {
		return err
	}
	return nil
}

// Shutdown gracefully shuts down the server.
func (s *Server) Shutdown(ctx context.Context) error {
	slog.Info("server shutting down")
	return s.httpServer.Shutdown(ctx)
}